drm: Update to Linux 4.12.14
* Initial Geminilake support

* Support improvements and model-specific bug fixes for Broxton, Haswell,
  Broadwell, Skylake and Kabylake hardware

* Various generic i915 bug fixes, including page flipping and memory
  corruption issues

* Minor radeon bug fixes

+++ sys/dev/agp/intel-gtt.c
@@ -1631,8 +1631,10 @@ intel_gtt_clear_range(u_int first_entry, u_int num_entries)
 	agp_intel_gtt_clear_range(intel_agp, first_entry, num_entries);
 }
 
-void intel_gtt_get(u64 *gtt_total, size_t *stolen_size,
-		   phys_addr_t *mappable_base, u64 *mappable_end)
+void intel_gtt_get(u64 *gtt_total,
+		   u32 *stolen_size,
+		   phys_addr_t *mappable_base,
+		   u64 *mappable_end)
 {
 	struct agp_info ainfo;
